refactor: Use CSS Modules to style OpenToWork.
