+++ components/Date.tsx
@@ -2,7 +2,7 @@ import { formatDate } from '../utils/formatDate';
 
 type Props = {
     date: string;
-    className: string;
+    className?: string;
 };
 
 export function Date({ date, ...props }: Props) {
