+++ components/OpenToWork/OpenToWork.module.scss
@@ -1,0 +1,42 @@
+@import '../../styles/breakpoints.module.scss';
+
+.wrapper {
+    background: var(--otw-background-color);
+
+    display: flex;
+    flex-direction: column;
+    gap: var(--spacing-3);
+
+    padding: var(--spacing-4) var(--spacing-3);
+
+    border-radius: var(--border-radius-3);
+    border-left: var(--border-width-2) solid var(--otw-border-color);
+
+    margin-left: calc(var(--spacing-5) * -1);
+    margin-right: calc(var(--spacing-5) * -1);
+
+    @media (max-width: $tablet-and-below) {
+        border: 1px solid var(--otw-border-color);
+        margin-left: 0;
+        margin-right: 0;
+    }
+}
+
+.heading {
+    color: var(--prose-otw-heading-color);
+    font-size: var(--font-size-base);
+}
+
+.details {
+    &[open] summary::before {
+        content: '🎯 ';
+    }
+}
+
+.summary {
+    list-style: none;
+
+    &::before {
+        content: '🔎 ';
+    }
+}
